split qcbor.h into four separate headers (#34)
This is to make it much easier to read the header files.

Backwards compatibility with qcbor.h is retained, but use of qcbor/qcbor_encode.h and qcbor/qcbor_decode.h is preferred now.
