+++ test/float_tests.c
@@ -11,7 +11,8 @@
  =============================================================================*/
 
 #include "float_tests.h"
-#include "qcbor.h"
+#include "qcbor/qcbor_encode.h"
+#include "qcbor/qcbor_decode.h"
 #include "half_to_double_from_rfc7049.h"
 #include <math.h> // For INFINITY and NAN and isnan()
