+++ test/qcbor_decode_tests.h
@@ -38,7 +38,8 @@ IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 /*
  Notes:
 
- - All the functions in qcbor.h are called once in the aggregation of all the tests below.
+ - All the functions in qcbor_decode.h are called once in the aggregation
+   of all the tests below.
 
  - All the types that are supported are given as input and parsed by these tests
